+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## vNEXT (not yet published)
 
+## v2.18.2
+
+### `@liveblocks/client`
+
+- Improve performance of undo/redo operations on large documents (thanks for the
+  contribution @rudi-c!)
+
+### `@liveblocks/react-tiptap`
+
+- Fix a performance regression introduced in 2.18.1
+
 ## v2.18.1
 
 ### `@liveblocks/react-ui`